# libs/backend/server/gateways/mcp/main/src/routes/mcp-servers.ts
import { Router, type Request } from "express";
import type { PrismaClient } from "@prisma/client";

import { addMcpServerCredential, createMcpServer, deleteMcpServer, deleteMcpServerCredential, getMcpServer, listMcpServerCredentials, listMcpServers, updateMcpServer } from "../core/mcp-servers.logic.js";
import { _RequireOrgAdmin } from "@opencrane/server/_infra/auth";
import type { McpServerCredentialInput, McpServerWriteRequest } from "./mcp-servers.types.js";

/**
 * CRUD router for the MCP server catalog.
 *
 * **Authorization (P0.5):** curating the catalogue is an org-admin action, so the
 * server lifecycle mutations (create / update / delete) are gated by `_RequireOrgAdmin`.
 * Reads stay open to any authenticated caller, and the per-server credential sub-routes
 * are intentionally NOT gated here — credential connect is a user action (spec §5.3),
 * to be designed with the P1 credential flow rather than locked to admins now.
 *
 * @param prisma - Prisma client used for persistence.
 * @returns Configured Express router.
 */
export function mcpServersRouter(prisma: PrismaClient): Router
{
  const router = Router();

  /** List all MCP servers with grants and credentials. */
  router.get("/", async function _listMcpServers(req, res)
  {
    res.json(await listMcpServers(prisma));
  });

  /** Get a single MCP server by identifier. */
  router.get("/:id", async function _getMcpServer(req, res)
  {
    const server = await getMcpServer(prisma, req.params.id);
    if (!server)
    {
      res.status(404).json({ error: "MCP server not found", code: "MCP_SERVER_NOT_FOUND" });
      return;
    }

    res.json(server);
  });

  /** Create a new MCP server plus generic grant rows for the compiler. Org-admin only. */
  router.post("/", _RequireOrgAdmin(), async function _createMcpServer(req, res)
  {
    const body = req.body as McpServerWriteRequest;
    res.status(201).json(await createMcpServer(prisma, body));
  });

  /** Update an MCP server and fully replace grants and credentials. Org-admin only. */
  router.put("/:id", _RequireOrgAdmin(), async function _updateMcpServer(req: Request<{ id: string }>, res)
  {
    const body = req.body as Partial<McpServerWriteRequest>;
    res.json(await updateMcpServer(prisma, req.params.id, body));
  });

  /** Delete an MCP server and its linked grant rows. Org-admin only. */
  router.delete("/:id", _RequireOrgAdmin(), async function _deleteMcpServer(req: Request<{ id: string }>, res)
  {
    res.json(await deleteMcpServer(prisma, req.params.id));
  });

  /** List the brokered credentials of an MCP server. */
  router.get("/:id/credentials", async function _listMcpServerCredentials(req, res)
  {
    const credentials = await listMcpServerCredentials(prisma, req.params.id);
    if (credentials === null)
    {
      res.status(404).json({ error: "MCP server not found", code: "MCP_SERVER_NOT_FOUND" });
      return;
    }

    res.json(credentials);
  });

  /** Add a single brokered credential to an MCP server (additive — grants untouched). */
  router.post("/:id/credentials", async function _addMcpServerCredential(req, res)
  {
    const credential = await addMcpServerCredential(prisma, req.params.id, req.body as McpServerCredentialInput);
    if (credential === null)
    {
      res.status(404).json({ error: "MCP server not found", code: "MCP_SERVER_NOT_FOUND" });
      return;
    }

    res.status(201).json(credential);
  });

  /** Remove a single brokered credential from an MCP server. */
  router.delete("/:id/credentials/:credentialId", async function _deleteMcpServerCredential(req, res)
  {
    const result = await deleteMcpServerCredential(prisma, req.params.id, req.params.credentialId);
    if (result === null)
    {
      res.status(404).json({ error: "MCP credential not found", code: "MCP_CREDENTIAL_NOT_FOUND" });
      return;
    }

    res.json(result);
  });

  return router;
}