# website/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenCrane Control Plane API",
    "version": "1.0.0",
    "description": "Multi-tenant AI agent platform management API.\n\n**Authentication**\n\n- *Human operators* — OIDC browser flow via `GET /auth/login` → `/auth/callback`. Session cookie is set server-side.\n- In-cluster workloads use short-lived, audience-bound projected service-account tokens at their dedicated internal trust boundaries.\n- Endpoints tagged *Auth* and *Meta* (`/auth/*`, `/openapi.json`) require no credentials."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API prefix"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "detail": {
            "type": "string",
            "description": "Optional extra context."
          }
        }
      },
...
      "McpServer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "transport": {
            "type": "string",
            "enum": [
              "streamable-http",
              "sse",
              "websocket"
            ]
          },
          "grants": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "credentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/McpServerCredential"
            }
          }
        }
      },
      "McpServerCredential": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable credential identifier."
          },
          "displayName": {
            "type": "string",
            "description": "Operator-facing label."
          }
        }
      },
      "McpCatalogServer": {
        "type": "object",
        "required": [
          "id"
        ],
        "description": "A catalogue server as exposed by the operator API (distinct from the registry McpServer). Every field beyond id is optional so the same shape serves the entitled user catalogue and the admin governance view.",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "glyph": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "single-user",
              "multi-user",
              "remote-oauth"
            ],
            "description": "Consumption shape; decides the credential-connect flow."
          },
          "approvalStatus": {
            "type": "string",
            "enum": [
              "pending-review",
              "approved",
              "published",
              "disabled"
            ],
            "description": "Governance lifecycle status."
          },
          "credentialSchema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialField"
            }
          },
          "entitlementSummary": {
            "type": "string",
            "description": "Human-readable summary of who is entitled (admin view)."
          }
        }
      },
      "CredentialField": {
        "type": "object",
        "required": [
          "key",
          "label",
          "required",
          "sensitive"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "Stable key the value is submitted under."
          },
          "label": {
            "type": "string",
            "description": "Human-readable field label."
          },
          "required": {
            "type": "boolean",
            "description": "Whether the field must be supplied."
          },
          "sensitive": {
            "type": "boolean",
            "description": "Whether the value is secret (masked, never echoed back)."
          },
          "placeholder": {
            "type": "string",
            "description": "Optional input placeholder."
          },
          "hint": {
            "type": "string",
            "description": "Optional helper hint."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/mcp-servers": {
      "get": {
        "operationId": "listMcpServers",
        "summary": "List all MCP servers with grants and credentials",
        "tags": [
          "MCP Servers"
        ],
        "responses": {
          "200": {
            "description": "MCP server list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpServer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMcpServer",
        "summary": "Create a new MCP server",
        "tags": [
          "MCP Servers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "endpoint",
                  "transport"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "transport": {
                    "type": "string"
                  },
                  "grants": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "credentials": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MCP server created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          }
        }
      }
    },
    "/mcp-servers/{id}": {
      "get": {
        "operationId": "getMcpServer",
        "summary": "Get a single MCP server by identifier",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "404": {
            "description": "MCP server not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateMcpServer",
        "summary": "Update an MCP server and fully replace grants and credentials",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP server updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMcpServer",
        "summary": "Delete an MCP server and its linked grant rows",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/mcp-servers/{id}/credentials": {
      "get": {
        "operationId": "listMcpServerCredentials",
        "summary": "List the brokered credentials of an MCP server",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpServerCredential"
                  }
                }
              }
            }
          },
          "404": {
            "description": "MCP server not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addMcpServerCredential",
        "summary": "Add a brokered credential to an MCP server (does not touch grants)",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "displayName"
                ],
                "properties": {
                  "displayName": {
                    "type": "string",
                    "description": "Operator-facing label."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credential added.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerCredential"
                }
              }
            }
          },
          "400": {
            "description": "Credential payload violates brokering-mode custody rules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "MCP server not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/mcp-servers/{id}/credentials/{credentialId}": {
      "delete": {
        "operationId": "deleteMcpServerCredential",
        "summary": "Remove a single brokered credential from an MCP server",
        "tags": [
          "MCP Servers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credentialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "MCP server or credential not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/servers": {
      "get": {
        "operationId": "listMcpGovernanceServers",
        "summary": "List every catalogue server regardless of status (org-admin governance view)",
        "tags": [
          "MCP Operator"
        ],
        "responses": {
          "200": {
            "description": "All catalogue servers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/McpCatalogServer"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an organisation admin.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}